pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["soy"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
SQLAlchemy = "^1.4.37"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-dotenv = "^0.20.0"
alembic = "^1.8.0"
mysqlclient = "^2.1.0"
PyYAML = "^6.0"
python-multipart = "^0.0.5"
mypy = "^0.960"
flake8 = "^4.0.1"
black = "^22.3.0"
pytest = "^7.1.2"
coverage = "^6.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
